# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: ERP MAFA Store — API
  version: "0.1.0"
  description: >
    Especificação de API gerada a partir dos controllers do Nest.js
    (apps/api). Ver convenções em docs/architecture/overview.md, seção 10.
servers:
  - url: /api/v1
...
paths:
...
  /sales:
    post:
      tags: [sales]
      summary: Cria venda em rascunho com itens (requer permission:manage_sales) — RN 10.10.1, não altera estoque
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [channel, items]
              properties:
                customerId: { type: string, format: uuid }
                channel: { type: string, enum: [presencial, whatsapp, instagram, catalogo, outro] }
                discount: { type: number, default: 0, description: Desconto geral da venda }
                items:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [productVariantId, quantity, unitPrice]
                    properties:
                      productVariantId: { type: string, format: uuid }
                      quantity: { type: number }
                      unitPrice: { type: number }
                      discount: { type: number, default: 0 }
      responses:
        "201":
          description: Criada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sale" }
    get:
      tags: [sales]
      summary: Lista vendas (requer permission:view_sales)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Sale" }
  /sales/{id}:
    patch:
      tags: [sales]
      summary: Edita os dados e substitui os itens de uma venda em rascunho (requer permission:manage_sales)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [channel, items]
              properties:
                customerId: { type: string, format: uuid }
                channel: { type: string, enum: [presencial, whatsapp, instagram, catalogo, outro] }
                discount: { type: number, default: 0 }
                items:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [productVariantId, quantity, unitPrice]
                    properties:
                      productVariantId: { type: string, format: uuid }
                      quantity: { type: number }
                      unitPrice: { type: number }
                      discount: { type: number, default: 0 }
      responses:
        "200":
          description: Rascunho atualizado; itens anteriores permanecem inativados no histórico
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sale" }
        "404":
          description: Venda não encontrada nesta empresa
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: Somente rascunhos podem ser editados
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    get:
      tags: [sales]
      summary: Detalhe da venda, com itens/pagamentos/devoluções (requer permission:view_sales)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sale" }
        "404":
          description: Venda não encontrada nesta empresa
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /sales/{id}/confirm:
    post:
      tags: [sales]
      summary: >
        Confirma a venda: congela o custo por item, baixa estoque via
        InventoryService.releaseStock() na mesma transação (TA-ARCH-003),
        registra os pagamentos (só à vista — a soma precisa bater exatamente
        com o total) e calcula CMV/lucro bruto.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [payments]
              properties:
                payments:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [paymentMethodId, amount]
                    properties:
                      paymentMethodId: { type: string, format: uuid }
                      amount: { type: number }
      responses:
        "201":
          description: Confirmada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sale" }
        "400":
          description: Estoque insuficiente ou soma dos pagamentos diferente do total
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: Venda não está em rascunho
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /sales/{id}/cancel:
    post:
      tags: [sales]
      summary: >
        Cancela a venda — permitido em draft (sem efeito) e confirmed
        (restaura o estoque ainda não devolvido e registra saída de estorno
        para cada entrada financeira original). Bloqueado se já
        returned/partially_returned/cancelled.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      responses:
        "201":
          description: Cancelada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sale" }
        "409":
          description: Status da venda não permite cancelamento
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /sales/{id}/return:
    post:
      tags: [sales]
      summary: >
        Devolução parcial ou total (RN 10.11). Item apto restaura estoque;
        item avariado só é registrado (não volta ao estoque disponível
        nesta fase). Recalcula receita líquida e CMV/lucro proporcionalmente.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason, items]
              properties:
                reason: { type: string }
                items:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [saleItemId, quantity, condition]
                    properties:
                      saleItemId: { type: string, format: uuid }
                      quantity: { type: number }
                      condition: { type: string, enum: [apt, damaged] }
      responses:
        "201":
          description: Devolução registrada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sale" }
        "400":
          description: Item inválido ou quantidade devolvida maior que a vendida
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: Status da venda não permite devolução
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /reporting/dashboard:
    get:
      tags: [reporting]
      summary: >
        Indicadores consolidados do período + comparação com o período anterior equivalente
        (requer permission:view_dashboard). O servidor compõe a resposta conforme as permissões
        de vendas, estoque, financeiro, custo e lucro do perfil; campos não autorizados são omitidos.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: from
          in: query
          required: true
          schema: { type: string, format: date }
        - name: to
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DashboardResponse" }
        "400":
          description: Período ausente ou inválido
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /reporting/sales:
    get:
      tags: [reporting]
      summary: Vendas do período com totais agregados (requer permission:view_reports)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: from
          in: query
          required: true
          schema: { type: string, format: date }
        - name: to
          in: query
          required: true
          schema: { type: string, format: date }
        - name: channel
          in: query
          schema: { type: string, enum: [presencial, whatsapp, instagram, catalogo, outro] }
        - name: status
          in: query
          schema: { type: string, enum: [draft, confirmed, cancelled, partially_returned, returned] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  period: { type: object, properties: { from: { type: string, format: date-time }, to: { type: string, format: date-time } } }
                  filters: { type: object }
                  totals:
                    type: object
                    properties:
                      grossRevenue: { type: string }
                      netRevenue: { type: string }
                      cmv: { type: string }
                      grossProfit: { type: string }
                      salesCount: { type: integer }
                  sales:
                    type: array
                    items: { $ref: "#/components/schemas/Sale" }
  /reporting/top-products:
    get:
      tags: [reporting]
      summary: Produtos mais vendidos/mais lucrativos no período (requer permission:view_reports)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: from
          in: query
          required: true
          schema: { type: string, format: date }
        - name: to
          in: query
          required: true
          schema: { type: string, format: date }
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
        - name: orderBy
          in: query
          schema: { type: string, enum: [quantity, profit], default: quantity }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  period: { type: object }
                  orderBy: { type: string }
                  products:
                    type: array
                    items:
                      type: object
                      properties:
                        productId: { type: string, format: uuid }
                        sku: { type: string }
                        name: { type: string }
                        quantitySold: { type: string }
                        revenue: { type: string }
                        profit: { type: string }
  /reporting/inventory-value:
    get:
      tags: [reporting]
      summary: >
        Valor de estoque atual, por produto e total (RN 11.10) — foto do momento, sem período
        (requer permission:view_reports).
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { type: string }
                  products:
                    type: array
                    items:
                      type: object
                      properties:
                        productId: { type: string, format: uuid }
                        sku: { type: string }
                        name: { type: string }
                        quantityAvailable: { type: string }
                        unitCost: { type: string }
                        value: { type: string }
  /reporting/dre:
    get:
      tags: [reporting]
      summary: DRE gerencial simplificada do período (RN 11.5/11.6/11.7, requer permission:view_reports)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: from
          in: query
          required: true
          schema: { type: string, format: date }
        - name: to
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DreResponse" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: false
      description: >
        TA-API-002 — chave opcional fornecida pelo cliente para evitar duplicar o efeito de um
        comando crítico em caso de reenvio (ex.: timeout de rede). Repetir a mesma chave com o
        mesmo payload devolve a resposta original; com payload diferente é tratado como uma nova
        operação, não como conflito.
      schema: { type: string, maxLength: 255 }
...
  schemas:
...
    Sale:
      type: object
      properties:
        id: { type: string, format: uuid }
        companyId: { type: string, format: uuid }
        customerId: { type: string, format: uuid, nullable: true }
        channel: { type: string, enum: [presencial, whatsapp, instagram, catalogo, outro] }
        status: { type: string, enum: [draft, confirmed, cancelled, partially_returned, returned] }
        subtotal: { type: string }
        discount: { type: string }
        total: { type: string }
        cmvCalculated: { type: string, nullable: true }
        grossProfitCalculated: { type: string, nullable: true }
        items:
          type: array
          items:
            type: object
            properties:
              id: { type: string, format: uuid }
              productVariantId: { type: string, format: uuid }
              quantity: { type: string }
              unitPrice: { type: string }
              discount: { type: string }
              unitCostAtSale: { type: string, nullable: true }
              quantityReturned: { type: string }
        payments:
          type: array
          items:
            type: object
            properties:
              paymentMethodId: { type: string, format: uuid }
              amount: { type: string }
              feeAmount: { type: string }
              netAmount: { type: string }
        returns:
          type: array
          items:
            type: object
            properties:
              id: { type: string, format: uuid }
              reason: { type: string }
              items:
                type: array
                items:
                  type: object
                  properties:
                    saleItemId: { type: string, format: uuid }
                    quantity: { type: string }
                    condition: { type: string, enum: [apt, damaged] }
...
    DashboardResponse:
      type: object
      description: >
        Resposta polimórfica por permissão. Métricas de vendas exigem view_sales; estoque exige
        view_inventory; valores financeiros exigem suas permissões específicas; CMV/lucro exigem
        view_cost/view_profit. Campos sem autorização são omitidos no servidor (RN 10.17.1/2).
      properties:
        period: { type: object, properties: { from: { type: string, format: date-time }, to: { type: string, format: date-time } } }
        revenueGross: { type: string }
        revenueNet: { type: string }
        salesCount: { type: integer }
        averageTicket: { type: string, nullable: true }
        expensesRealized: { type: string }
        productsCount: { type: integer }
        lowStockCount: { type: integer }
        inventoryValue: { type: string }
        receivablesOpen: { type: string }
        overdueReceivablesCount: { type: integer }
        payablesOpen: { type: string }
        overduePayablesCount: { type: integer }
        cashBalance: { type: string }
        cmv: { type: string, description: "Só com view_cost" }
        grossProfit: { type: string, description: "Só com view_profit" }
        netProfitEstimated: { type: string, description: "Só com view_profit" }
        margin: { type: string, nullable: true, description: "Só com view_profit" }
        comparison:
          type: object
          properties:
            revenueNetChangePercent: { type: string, nullable: true }
            salesCountChangePercent: { type: string, nullable: true }
            grossProfitChangePercent: { type: string, nullable: true, description: "Só com view_profit" }
            netProfitChangePercent: { type: string, nullable: true, description: "Só com view_profit" }
        salesTrend:
          type: array
          description: "Só com view_sales; série diária das vendas válidas do período."
          items:
            type: object
            properties:
              date: { type: string, format: date }
              revenue: { type: string }
              salesCount: { type: integer }
        topProducts:
          type: array
          description: "Só com view_sales; profit é omitido sem view_profit."
          items:
            type: object
            properties:
              productId: { type: string, format: uuid }
              sku: { type: string }
              name: { type: string }
              quantitySold: { type: string }
              revenue: { type: string }
              profit: { type: string, description: "Só com view_profit" }
    DreResponse:
      type: object
      properties:
        period: { type: object }
        grossRevenue: { type: string }
        discountsAndReturns: { type: string }
        netRevenue: { type: string }
        cmv: { type: string }
        grossProfit: { type: string }
        paymentFees: { type: string }
        expenses: { type: string }
        netProfit: { type: string }
...
    ErrorResponse:
      type: object
      description: Formato padrão TA-API-001 (docs/architecture/overview.md, secao 10.2)
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            field: { type: string }
            details: { type: object }